Use float between [0, 1] for Color

+++ include/mng/color.h
@@ -1,17 +1,15 @@
 #ifndef MNG_COLOR_H
 #define MNG_COLOR_H
 
-#include <stdint.h>
-
 #ifdef __cplusplus
 extern "C" {
 #endif
 
 typedef struct Color {
-    uint8_t red;
-    uint8_t green;
-    uint8_t blue;
-    uint8_t alpha;
+    float red;
+    float green;
+    float blue;
+    float alpha;
 } Color;
 
 #ifdef __cplusplus
